requirements.txt
telegram==0.0.1
requests==2.25.1
python-dotenv==0.17.1

